Amendment 11 2061-12-19
Objectives

Election of Vice President by Electoral College consisting of members of both Houses of Parliament, instead of election by a Joint Sitting of Parliament. Indemnify the President and Vice President Election procedure from challenge on grounds of existence of any vacancies in the electoral college.

Amendments

Amend articles 66 and 71.[15]

Closes #15

Signed-off-by: sushmit1 <sushmit718@gmail.com>